requirements.txt
langchain
pypdf
sentence-transformers
faiss-cpu
llama-cpp-python
streamlit
